+++ tests/Aql/ExecutionNode/MockTypedNode.cpp
